start on SingleRouter
