+++ packages/houdini-react/src/runtime/routing/components/Link.tsx
@@ -1,6 +1,6 @@
 import React, { useTransition } from 'react'
 
-import { useRouterContext } from './Router'
+import { useRouterContext } from './SingleRouter'
 
 export function Link({
 	href,
