+++ packages/houdini-react/src/runtime/routing/index.ts
@@ -1,1 +1,2 @@
 export * from './components'
+export type { RouterManifest, RouterPageManifest } from './lib/types'
